♻️refactor: add intersection observer to carousels
